Pin Buildkite test images
This ensures that we use variable in our Makefile to track the buildkite
test images we use. Note that we only pin to an image we can track (the
linter has no versioning yet). This ensures that we don't get hit by the
recently broken bats-mock update which comes as part of v3.0.0.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>